refactor: move canvas renderer into builder
There is no need for canvas renderer in react-sdk.
Moved into builder and slightly cleaned.
