+++ packages/react-sdk/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./remix";
 export * from "./css/index";
-export * from "./tree/index";
 export * from "./core-components";
 export * from "./components/components-utils";
 export { PropMeta } from "./prop-meta";
