--- packages/react-sdk/src/tree/index.ts
@@ -1,1 +1,0 @@
-export * from "./create-elements-tree";
